+++ src/APE.PostgreSQL.Teamwork.Model/Setting/DatabaseSetting.cs
@@ -5,6 +5,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 using APE.PostgreSQL.Teamwork.Model.Setting;
+using APE.PostgreSQL.Teamwork.Model.Templates;
 
 namespace APE.PostgreSQL.Teamwork.Model
 {
@@ -13,6 +14,12 @@ namespace APE.PostgreSQL.Teamwork.Model
     /// </summary>
     public class DatabaseSetting
     {
+        public static readonly List<string> DefaultIgnoredSchemas = new List<string>()
+        {
+            SQLTemplates.PostgreSQLTeamworkSchemaName,
+            "APE.PostgreSQL.Test.Runner",
+        };
+
         /// <summary>
         /// Creates a new <see cref="DatabaseSetting"/>.
         /// </summary>
@@ -47,6 +54,11 @@ namespace APE.PostgreSQL.Teamwork.Model
         /// Gets or sets the path to the diffs and dumps.
         /// </summary>
         public string Path { get; set; }
+
+        /// <summary>
+        /// Gets or sets the ignored schemas.
+        /// </summary>
+        public List<string> IgnoredSchemas { get; set; } = new List<string>();
 
         /// <summary>
         /// Removes the database with the given id.