+++ src/APE.PostgreSQL.Teamwork.ViewModel/Exceptions/TeamworkTestException.cs
@@ -1,8 +1,6 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
+﻿// <copyright file="TeamworkTestException.cs" company="APE Engineering GmbH">Copyright (c) APE Engineering GmbH. All rights reserved.</copyright>
+
+using System;
 
 namespace APE.PostgreSQL.Teamwork.ViewModel.Exceptions
 {